Extract VS Code test files into `./test/vscode`
To prevent accidentally building and shipping the test files with the
extension, which I think would have happened now.
