+++ .gitignore
@@ -1,6 +1,3 @@
 node_modules
 /vscode/out
-/vscode/.vscode-test
-/vscode/webview-ui/node_modules
-/vscode/webview-ui/dist
 /build
