+++ Makefile
@@ -1,11 +1,16 @@
 NPM = npm
 CODE = code
 
+all: webview vscode vscode-test
+
 vscode: .always
 	cd vscode && $(NPM) run install:all
-	cd vscode && $(NPM) run build
 	cd vscode && $(NPM) run lint
-	cd vscode && $(NPM) test
+	cd vscode && $(NPM) run build
+
+vscode-test: .always
+	cd test/vscode && $(NPM) ci
+	cd test/vscode && $(NPM) test
 
 vscode-open: .always
 	$(CODE) --extensionDevelopmentPath="$(PWD)/vscode"
