+++ test/vscode/fixtures/test-schema.json
