+++ test/vscode/package.json
@@ -1,0 +1,24 @@
+{
+  "name": "@sourcemeta/studio-test-vscode",
+  "version": "0.0.0",
+  "private": true,
+  "description": "Tests for Sourcemeta Studio VSCode extension",
+  "scripts": {
+    "compile": "tsc -p ./",
+    "pretest": "npm run compile",
+    "test": "vscode-test",
+    "test:headed": "vscode-test --headless=false",
+    "test:watch": "vscode-test --watch"
+  },
+  "devDependencies": {
+    "@types/chai": "^5.2.3",
+    "@types/mocha": "^10.0.10",
+    "@types/node": "18.x",
+    "@types/vscode": "^1.74.0",
+    "@vscode/test-cli": "^0.0.12",
+    "@vscode/test-electron": "^2.5.2",
+    "chai": "^6.2.0",
+    "mocha": "^11.7.4",
+    "typescript": "^5.3.0"
+  }
+}
