+++ test/vscode/tsconfig.json
@@ -2,7 +2,7 @@
   "compilerOptions": {
     "module": "commonjs",
     "target": "ES2020",
-    "outDir": "out",
+    "outDir": "../../build/test/vscode",
     "lib": ["ES2020"],
     "sourceMap": true,
     "strict": true,
@@ -14,8 +14,13 @@
     "noUnusedLocals": true,
     "noUnusedParameters": true,
     "noImplicitReturns": true,
-    "noFallthroughCasesInSwitch": true
+    "noFallthroughCasesInSwitch": true,
+    "types": ["mocha", "node", "@types/vscode"]
   },
-  "include": ["src/**/*", "../shared/**/*"],
-  "exclude": ["node_modules", "out"]
+  "include": [
+    "**/*",
+    "../../vscode/src/**/*",
+    "../../shared/**/*"
+  ],
+  "exclude": ["node_modules"]
 }
