+++ vscode/package.json
@@ -50,27 +50,17 @@
     "watch": "tsc -watch -p ./",
     "build:webview": "cd ../webview && npm run build",
     "build": "npm run compile && npm run build:webview",
-    "pretest": "npm run compile",
-    "test": "vscode-test",
-    "test:headed": "vscode-test --headless=false",
-    "test:watch": "vscode-test --watch",
     "lint": "eslint src --ext ts"
   },
   "dependencies": {
     "@sourcemeta/jsonschema": "^12.2.0"
   },
   "devDependencies": {
-    "@types/chai": "^5.2.3",
-    "@types/mocha": "^10.0.10",
     "@types/node": "18.x",
     "@types/vscode": "^1.74.0",
     "@typescript-eslint/eslint-plugin": "^8.46.2",
     "@typescript-eslint/parser": "^8.46.2",
-    "@vscode/test-cli": "^0.0.12",
-    "@vscode/test-electron": "^2.5.2",
-    "chai": "^6.2.0",
     "eslint": "^9.38.0",
-    "mocha": "^11.7.4",
     "typescript": "^5.3.0"
   }
 }
